refactor: move ValidateRayClusterSpec to validation.go and its unit test to validation_test.go
