+++ ray-operator/controllers/ray/utils/validation.go
@@ -2,10 +2,10 @@ package utils
 
 import (
 	errstd "errors"
-
-	"k8s.io/apimachinery/pkg/api/meta"
-
+	"fmt"
 	rayv1 "github.com/ray-project/kuberay/ray-operator/apis/ray/v1"
+	"github.com/ray-project/kuberay/ray-operator/pkg/features"
+	"k8s.io/apimachinery/pkg/api/meta"
 )
 
 func ValidateRayClusterStatus(instance *rayv1.RayCluster) error {
@@ -13,6 +13,73 @@ func ValidateRayClusterStatus(instance *rayv1.RayCluster) error {
 	suspended := meta.IsStatusConditionTrue(instance.Status.Conditions, string(rayv1.RayClusterSuspended))
 	if suspending && suspended {
 		return errstd.New("invalid RayCluster State: rayv1.RayClusterSuspending and rayv1.RayClusterSuspended conditions should not be both true")
+	}
+	return nil
+}
+
+// Validation for invalid Ray Cluster configurations.
+func ValidateRayClusterSpec(instance *rayv1.RayCluster) error {
+	if len(instance.Spec.HeadGroupSpec.Template.Spec.Containers) == 0 {
+		return fmt.Errorf("headGroupSpec should have at least one container")
+	}
+
+	for _, workerGroup := range instance.Spec.WorkerGroupSpecs {
+		if len(workerGroup.Template.Spec.Containers) == 0 {
+			return fmt.Errorf("workerGroupSpec should have at least one container")
+		}
+	}
+
+	if instance.Annotations[RayFTEnabledAnnotationKey] != "" && instance.Spec.GcsFaultToleranceOptions != nil {
+		return fmt.Errorf("%s annotation and GcsFaultToleranceOptions are both set. "+
+			"Please use only GcsFaultToleranceOptions to configure GCS fault tolerance", RayFTEnabledAnnotationKey)
+	}
+
+	if !IsGCSFaultToleranceEnabled(*instance) {
+		if EnvVarExists(RAY_REDIS_ADDRESS, instance.Spec.HeadGroupSpec.Template.Spec.Containers[RayContainerIndex].Env) {
+			return fmt.Errorf("%s is set which implicitly enables GCS fault tolerance, "+
+				"but GcsFaultToleranceOptions is not set. Please set GcsFaultToleranceOptions "+
+				"to enable GCS fault tolerance", RAY_REDIS_ADDRESS)
+		}
+	}
+
+	if instance.Spec.GcsFaultToleranceOptions != nil {
+		if redisPassword := instance.Spec.HeadGroupSpec.RayStartParams["redis-password"]; redisPassword != "" {
+			return fmt.Errorf("cannot set `redis-password` in rayStartParams when " +
+				"GcsFaultToleranceOptions is enabled - use GcsFaultToleranceOptions.RedisPassword instead")
+		}
+
+		headContainer := instance.Spec.HeadGroupSpec.Template.Spec.Containers[RayContainerIndex]
+		if EnvVarExists(REDIS_PASSWORD, headContainer.Env) {
+			return fmt.Errorf("cannot set `REDIS_PASSWORD` env var in head Pod when " +
+				"GcsFaultToleranceOptions is enabled - use GcsFaultToleranceOptions.RedisPassword instead")
+		}
+
+		if EnvVarExists(RAY_REDIS_ADDRESS, headContainer.Env) {
+			return fmt.Errorf("cannot set `RAY_REDIS_ADDRESS` env var in head Pod when " +
+				"GcsFaultToleranceOptions is enabled - use GcsFaultToleranceOptions.RedisAddress instead")
+		}
+
+		if instance.Annotations[RayExternalStorageNSAnnotationKey] != "" {
+			return fmt.Errorf("cannot set `ray.io/external-storage-namespace` annotation when " +
+				"GcsFaultToleranceOptions is enabled - use GcsFaultToleranceOptions.ExternalStorageNamespace instead")
+		}
+	}
+
+	if !features.Enabled(features.RayJobDeletionPolicy) {
+		for _, workerGroup := range instance.Spec.WorkerGroupSpecs {
+			if workerGroup.Suspend != nil && *workerGroup.Suspend {
+				return fmt.Errorf("suspending worker groups is currently available when the RayJobDeletionPolicy feature gate is enabled")
+			}
+		}
+	}
+
+	if IsAutoscalingEnabled(instance) {
+		for _, workerGroup := range instance.Spec.WorkerGroupSpecs {
+			if workerGroup.Suspend != nil && *workerGroup.Suspend {
+				// TODO (rueian): This can be supported in future Ray. We should check the RayVersion once we know the version.
+				return fmt.Errorf("suspending worker groups is not currently supported with Autoscaler enabled")
+			}
+		}
 	}
 	return nil
 }
